# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Comet API",
    "version": "1.0.0",
    "description": "API documentation for Comet: The Speed of Light Comic Reader."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API"
    }
  ],
  "components": {
    "schemas": {
...
      "CollectionDTO": {
        "type": "object",
        "required": [
          "id",
          "name",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "_count": {
            "type": "object",
            "properties": {
              "items": {
                "type": "integer"
              }
            }
          }
        }
      },
      "CollectionCreateSchema": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50
          },
          "description": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
...
      "BookmarkDTO": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "comicId",
          "page",
          "note",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "comicId": {
            "type": "string"
          },
          "page": {
            "type": "integer"
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "BookmarkCreateSchema": {
        "type": "object",
        "required": [
          "comicId",
          "page"
        ],
        "properties": {
          "comicId": {
            "type": "string"
          },
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "note": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ComicVineMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "issue_number": {
            "type": "string"
          },
          "cover_date": {
            "type": "string"
          },
          "image": {
            "type": "object",
            "properties": {
              "icon_url": {
                "type": "string"
              },
              "medium_url": {
                "type": "string"
              },
              "screen_url": {
                "type": "string"
              },
              "super_url": {
                "type": "string"
              },
              "thumb_url": {
                "type": "string"
              },
              "tiny_url": {
                "type": "string"
              },
              "original_url": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/collections": {
      "get": {
        "summary": "List collections",
        "tags": [
          "Collections"
        ],
        "responses": {
          "200": {
            "description": "List of user collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "collections": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CollectionDTO"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_collections"
      },
      "post": {
        "summary": "Create collection",
        "tags": [
          "Collections"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection created successfully"
          },
          "400": {
            "description": "Validation error or duplicate name"
          }
        },
        "operationId": "post_collections"
      }
    },
    "/bookmarks": {
      "get": {
        "summary": "List bookmarks",
        "tags": [
          "Bookmarks"
        ],
        "responses": {
          "200": {
            "description": "List of bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookmarkDTO"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_bookmarks"
      },
      "post": {
        "summary": "Create a bookmark",
        "tags": [
          "Bookmarks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreateSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bookmark created"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "post_bookmarks"
      },
      "delete": {
        "summary": "Delete a bookmark",
        "tags": [
          "Bookmarks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bookmark deleted"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "delete_bookmarks"
      }
    },
    "/collections/{id}": {
      "get": {
        "summary": "Get collection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "get_collections_id"
      },
      "patch": {
        "summary": "Update collection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionCreateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collection updated"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "patch_collections_id"
      },
      "delete": {
        "summary": "Delete collection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection deleted"
          },
          "404": {
            "description": "Not found"
          }
        },
        "operationId": "delete_collections_id"
      }
    },
    "/collections/{id}/items": {
      "post": {
        "summary": "Add comic to collection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "comicId"
                ],
                "properties": {
                  "comicId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Added successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "post_collections_id_items"
      },
      "delete": {
        "summary": "Remove comic from collection",
        "tags": [
          "Collections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "comicId"
                ],
                "properties": {
                  "comicId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Removed successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "delete_collections_id_items"
      }
    },
...
    "/comics/{id}/enrich": {
      "get": {
        "summary": "Fetch ComicVine metadata",
        "tags": [
          "Comics"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Enrichment result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComicVineMetadata"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "operationId": "get_comics_id_enrich"
      }
    },
...
  }
}